requirements.txt
click~=7.1.2
pydantic~=1.7.3
PyYAML~=5.3.1
